Hacky way to persist hidden attribute in block of Set
To persist hidden attribute in state is tricky, while one can use a
`Computed` attribute to do that. However, if the property is inside a
nested block whose containing type is `Set`, things will get more
complicated. Since once something changed, the hash of the new schema
changed, and the `d` you got in callback is only be able to get the attr
for the new state, not the previous state. The good news is that the
original state is there, in `d`. So we introduced a hacky way to get it
in `getHiddenAttribute` (the reason why using `unsafe.Pointer` here is
because the `state` attribute is private attribute, reflect is not
capable to hanlde that case)